fix unused-variables

+++ src/mongocxx/test/gridfs/bucket.cpp
@@ -119,8 +119,7 @@ void validate_gridfs_file(
 
     REQUIRE(files_doc->view()["_id"].get_oid() == id.get_oid());
     REQUIRE(files_doc->view()["length"].get_int64().value == expected_length);
-    REQUIRE(static_cast<std::size_t>(files_doc->view()["chunkSize"].get_int32().value) ==
-            expected_chunk_size);
+    REQUIRE(files_doc->view()["chunkSize"].get_int32().value == expected_chunk_size);
     REQUIRE(files_doc->view()["filename"].get_utf8().value ==
             stdx::string_view(expected_file_name));
