+++ src/mongocxx/test/index_view.cpp
@@ -344,9 +344,9 @@ TEST_CASE("drop_all", "[index_view]") {
         bsoncxx::document::view result_view = result.view();
 
         auto cursor1 = indexes.list();
-        REQUIRE(std::distance(cursor1.begin(), cursor1.end()) == models.size() + 1);
-        REQUIRE((result_view["numIndexesAfter"].get_int32() -
-                 result_view["numIndexesBefore"].get_int32()) == models.size());
+        REQUIRE((unsigned)std::distance(cursor1.begin(), cursor1.end()) == models.size() + 1);
+        REQUIRE((unsigned)(result_view["numIndexesAfter"].get_int32() -
+                           result_view["numIndexesBefore"].get_int32()) == models.size());
 
         indexes.drop_all();
         auto cursor2 = indexes.list();
@@ -365,9 +365,9 @@ TEST_CASE("drop_all", "[index_view]") {
         bsoncxx::document::view result_view = result.view();
 
         auto cursor1 = indexes.list();
-        REQUIRE(std::distance(cursor1.begin(), cursor1.end()) == models.size() + 1);
-        REQUIRE((result_view["numIndexesAfter"].get_int32() -
-                 result_view["numIndexesBefore"].get_int32()) == models.size());
+        REQUIRE((unsigned)std::distance(cursor1.begin(), cursor1.end()) == models.size() + 1);
+        REQUIRE((unsigned)(result_view["numIndexesAfter"].get_int32() -
+                           result_view["numIndexesBefore"].get_int32()) == models.size());
 
         options::index_view options;
         options.max_time(std::chrono::milliseconds(1));
